feat/administrator

+++ simple-dpgf-api/src/main/java/com/simpledpgfapi/user/service/InvitationService.java
@@ -32,7 +32,7 @@ public class InvitationService {
     @Autowired
     private UserService userService;
     @Autowired
-    private UserAuthenticationService userAuthenticationService;
+    private AuthenticationService authenticationService;
     @Autowired
     private OrganizationRepository organizationRepository;
     @Autowired
@@ -44,12 +44,15 @@ public class InvitationService {
         User currentUser = userService.getCurrentAuthenticatedUser();
 
         Invitation existingInvitation = invitationRepository.findByEmailReceiver(invitationCreationDto.getEmailReceiver());
-        if(existingInvitation.getInvitationStatus() == InvitationStatusEnum.CONSUMED) {
-            throw new HttpException(HttpStatus.BAD_REQUEST, InvitationErrorCodes.INVITATION_CONSUMED);
-        }
 
-        if(existingInvitation.getInvitationStatus() == InvitationStatusEnum.PENDING) {
-            invitationRepository.delete(existingInvitation);
+        if(existingInvitation != null) {
+            if(existingInvitation.getInvitationStatus() == InvitationStatusEnum.CONSUMED) {
+                throw new HttpException(HttpStatus.BAD_REQUEST, InvitationErrorCodes.INVITATION_CONSUMED);
+            }
+
+            if(existingInvitation.getInvitationStatus() == InvitationStatusEnum.PENDING) {
+                invitationRepository.delete(existingInvitation);
+            }
         }
 
         Invitation invitation = createProjectOwnerInvitation(currentUser.getEmail(), invitationCreationDto, currentUser.getOrganizationId());
@@ -87,7 +90,7 @@ public class InvitationService {
         invitationRepository.save(invitation);
 
         userCreationDto.setOrganization(organizationCreationDto);
-        return userAuthenticationService.createUser(userCreationDto, invitation.getRole());
+        return authenticationService.createUser(userCreationDto, invitation.getRole());
     }
 
     private Invitation createProjectOwnerInvitation(String emailSender, InvitationCreationDto invitationCreationDto, ObjectId organizationId) {
